use annotation for json document editor as well
